requirements.txt
Django
MySQL-python
django-rest-multiple-models
djangorestframework
markdown
django-filter